wip: start building duplex encryption specification for Poseidon
