+++ manta-parameters/Cargo.toml
@@ -32,7 +32,7 @@ download = ["anyhow", "attohttpc", "std"]
 std = ["anyhow?/std"]
 
 [dependencies]
-anyhow = { version = "1.0.57", optional = true, default-features = false }
+anyhow = { version = "1.0.58", optional = true, default-features = false }
 attohttpc = { version = "0.19.1", optional = true }
 blake3 = { version = "1.3.1", default-features = false }
 
@@ -44,7 +44,7 @@ tempfile = { version = "3.3.0", default-features = false }
 walkdir = { version = "2.3.2", default-features = false }
 
 [build-dependencies]
-anyhow = { version = "1.0.57", default-features = false, features = ["std"] }
+anyhow = { version = "1.0.58", default-features = false, features = ["std"] }
 blake3 = { version = "1.3.1", default-features = false, features = ["std"] }
 hex = { version = "0.4.3", default-features = false, features = ["std"] }
 walkdir = { version = "2.3.2", default-features = false }
