+++ manta-pay/src/crypto/poseidon/mod.rs
@@ -22,25 +22,27 @@ use crate::crypto::poseidon::{
 use alloc::{boxed::Box, vec::Vec};
 use core::{fmt::Debug, hash::Hash, iter, mem, slice};
 use manta_crypto::{
-    hash::ArrayHashFunction,
     permutation::PseudorandomPermutation,
-    rand::{Rand, RngCore, Sample},
+    rand::{RngCore, Sample},
 };
-use manta_util::{
-    codec::{Decode, DecodeError, Encode, Read, Write},
-    vec::VecExt,
-};
+use manta_util::codec::{Decode, DecodeError, Encode, Read, Write};
 
 #[cfg(feature = "serde")]
 use manta_util::serde::{Deserialize, Serialize};
 
 pub mod compat;
 pub mod constants;
+pub mod encryption;
+pub mod hash;
 pub mod lfsr;
 pub mod matrix;
 pub mod mds;
 pub mod preprocessing;
 pub mod round_constants;
+
+#[cfg(feature = "arkworks")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "arkworks")))]
+pub mod arkworks;
 
 /// Field Element
 pub trait Field {
@@ -429,382 +431,5 @@ where
                 .to_row_major()
                 .into_boxed_slice(),
         }
-    }
-}
-
-/// Poseidon Hasher
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(
-        bound(
-            deserialize = "Permutation<S, COM>: Deserialize<'de>, S::Field: Deserialize<'de>",
-            serialize = "Permutation<S, COM>: Serialize, S::Field: Serialize"
-        ),
-        crate = "manta_util::serde",
-        deny_unknown_fields
-    )
-)]
-#[derive(derivative::Derivative)]
-#[derivative(
-    Clone(bound = "Permutation<S, COM>: Clone, S::Field: Clone"),
-    Debug(bound = "Permutation<S, COM>: Debug, S::Field: Debug"),
-    Eq(bound = "Permutation<S, COM>: Eq, S::Field: Eq"),
-    Hash(bound = "Permutation<S, COM>: Hash, S::Field: Hash"),
-    PartialEq(bound = "Permutation<S, COM>: PartialEq, S::Field: PartialEq")
-)]
-pub struct Hasher<S, const ARITY: usize, COM = ()>
-where
-    S: Specification<COM>,
-{
-    /// Poseidon Permutation
-    permutation: Permutation<S, COM>,
-
-    /// Domain Tag
-    domain_tag: S::Field,
-}
-
-impl<S, const ARITY: usize, COM> Hasher<S, ARITY, COM>
-where
-    S: Specification<COM>,
-{
-    /// Builds a new [`Hasher`] over `permutation` and `domain_tag`.
-    #[inline]
-    pub fn new(permutation: Permutation<S, COM>, domain_tag: S::Field) -> Self {
-        assert_eq!(ARITY + 1, S::WIDTH);
-        Self {
-            permutation,
-            domain_tag,
-        }
-    }
-}
-
-impl<S, const ARITY: usize, COM> ArrayHashFunction<ARITY, COM> for Hasher<S, ARITY, COM>
-where
-    S: Specification<COM>,
-{
-    type Input = S::Field;
-    type Output = S::Field;
-
-    #[inline]
-    fn hash(&self, input: [&Self::Input; ARITY], compiler: &mut COM) -> Self::Output {
-        let mut state = self.permutation.first_round_with_domain_tag_unchecked(
-            &self.domain_tag,
-            input,
-            compiler,
-        );
-        self.permutation
-            .permute_without_first_round(&mut state, compiler);
-        state.0.into_vec().take_first()
-    }
-}
-
-impl<S, const ARITY: usize, COM> Decode for Hasher<S, ARITY, COM>
-where
-    S: Specification<COM>,
-    S::Field: Decode,
-    S::ParameterField: Decode<Error = <S::Field as Decode>::Error>,
-{
-    type Error = <S::Field as Decode>::Error;
-
-    #[inline]
-    fn decode<R>(mut reader: R) -> Result<Self, DecodeError<R::Error, Self::Error>>
-    where
-        R: Read,
-    {
-        Ok(Self::new(
-            Decode::decode(&mut reader)?,
-            Decode::decode(&mut reader)?,
-        ))
-    }
-}
-
-impl<S, const ARITY: usize, COM> Encode for Hasher<S, ARITY, COM>
-where
-    S: Specification<COM>,
-    S::Field: Encode,
-    S::ParameterField: Encode,
-{
-    #[inline]
-    fn encode<W>(&self, mut writer: W) -> Result<(), W::Error>
-    where
-        W: Write,
-    {
-        self.permutation.encode(&mut writer)?;
-        self.domain_tag.encode(&mut writer)?;
-        Ok(())
-    }
-}
-
-impl<D, S, const ARITY: usize, COM> Sample<D> for Hasher<S, ARITY, COM>
-where
-    D: Clone,
-    S: Specification<COM>,
-    S::Field: Sample<D>,
-    S::ParameterField: Field + FieldGeneration + PartialEq + Sample<D>,
-{
-    /// Samples random Poseidon parameters.
-    #[inline]
-    fn sample<R>(distribution: D, rng: &mut R) -> Self
-    where
-        R: RngCore + ?Sized,
-    {
-        // FIXME: Use a proper domain tag sampling method.
-        Self::new(rng.sample(distribution.clone()), rng.sample(distribution))
-    }
-}
-
-/// Arkworks Backend.
-#[cfg(feature = "arkworks")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "arkworks")))]
-pub mod arkworks {
-    use crate::crypto::{
-        constraint::arkworks::{Fp, FpVar, R1CS},
-        poseidon::FieldGeneration,
-    };
-    use ark_ff::{BigInteger, Field, FpParameters, PrimeField};
-    use ark_r1cs_std::fields::FieldVar;
-    use manta_crypto::constraint::Constant;
-
-    /// Compiler Type.
-    type Compiler<S> = R1CS<<S as Specification>::Field>;
-
-    /// Poseidon Permutation Specification.
-    pub trait Specification {
-        /// Field Type
-        type Field: PrimeField;
-
-        /// Width of the Permutation
-        ///
-        /// This number is the total number `t` of field elements in the state which is `F^t`.
-        const WIDTH: usize;
-
-        /// Number of Partial Rounds
-        const PARTIAL_ROUNDS: usize;
-
-        /// Number of Full Rounds
-        ///
-        /// The total number of full rounds in the Poseidon permutation, including the first set of
-        /// full rounds and then the second set after the partial rounds.
-        const FULL_ROUNDS: usize;
-
-        /// S-BOX Exponenet
-        const SBOX_EXPONENT: u64;
-    }
-
-    impl<F> super::Field for Fp<F>
-    where
-        F: PrimeField,
-    {
-        #[inline]
-        fn zero() -> Self {
-            Self(F::zero())
-        }
-
-        #[inline]
-        fn is_zero(&self) -> bool {
-            self.0 == F::zero()
-        }
-
-        #[inline]
-        fn one() -> Self {
-            Self(F::one())
-        }
-
-        #[inline]
-        fn add(lhs: &Self, rhs: &Self) -> Self {
-            Self(lhs.0 + rhs.0)
-        }
-
-        #[inline]
-        fn add_assign(&mut self, rhs: &Self) {
-            self.0 += rhs.0;
-        }
-
-        #[inline]
-        fn sub(lhs: &Self, rhs: &Self) -> Self {
-            Self(lhs.0 - rhs.0)
-        }
-
-        #[inline]
-        fn mul(lhs: &Self, rhs: &Self) -> Self {
-            Self(lhs.0 * rhs.0)
-        }
-
-        #[inline]
-        fn inverse(&self) -> Option<Self> {
-            self.0.inverse().map(Self)
-        }
-    }
-
-    impl<F> FieldGeneration for Fp<F>
-    where
-        F: PrimeField,
-    {
-        const MODULUS_BITS: usize = F::Params::MODULUS_BITS as usize;
-
-        #[inline]
-        fn try_from_bits_be(bits: &[bool]) -> Option<Self> {
-            F::from_repr(F::BigInt::from_bits_be(bits)).map(Self)
-        }
-
-        #[inline]
-        fn from_u64(elem: u64) -> Self {
-            Self(F::from(elem))
-        }
-    }
-
-    impl<S> super::Specification for S
-    where
-        S: Specification,
-    {
-        type Field = Fp<S::Field>;
-        type ParameterField = Fp<S::Field>;
-
-        const WIDTH: usize = S::WIDTH;
-        const FULL_ROUNDS: usize = S::FULL_ROUNDS;
-        const PARTIAL_ROUNDS: usize = S::PARTIAL_ROUNDS;
-
-        #[inline]
-        fn add(lhs: &Self::Field, rhs: &Self::Field, _: &mut ()) -> Self::Field {
-            Fp(lhs.0 + rhs.0)
-        }
-
-        #[inline]
-        fn add_const(lhs: &Self::Field, rhs: &Self::ParameterField, _: &mut ()) -> Self::Field {
-            Fp(lhs.0 + rhs.0)
-        }
-
-        #[inline]
-        fn mul(lhs: &Self::Field, rhs: &Self::Field, _: &mut ()) -> Self::Field {
-            Fp(lhs.0 * rhs.0)
-        }
-
-        #[inline]
-        fn mul_const(lhs: &Self::Field, rhs: &Self::ParameterField, _: &mut ()) -> Self::Field {
-            Fp(lhs.0 * rhs.0)
-        }
-
-        #[inline]
-        fn add_assign(lhs: &mut Self::Field, rhs: &Self::Field, _: &mut ()) {
-            lhs.0 += rhs.0;
-        }
-
-        #[inline]
-        fn add_const_assign(lhs: &mut Self::Field, rhs: &Self::ParameterField, _: &mut ()) {
-            lhs.0 += rhs.0;
-        }
-
-        #[inline]
-        fn apply_sbox(point: &mut Self::Field, _: &mut ()) {
-            point.0 = point.0.pow(&[Self::SBOX_EXPONENT, 0, 0, 0]);
-        }
-    }
-
-    impl<S> super::Specification<Compiler<S>> for S
-    where
-        S: Specification,
-    {
-        type Field = FpVar<S::Field>;
-        type ParameterField = Fp<S::Field>;
-
-        const WIDTH: usize = S::WIDTH;
-        const FULL_ROUNDS: usize = S::FULL_ROUNDS;
-        const PARTIAL_ROUNDS: usize = S::PARTIAL_ROUNDS;
-
-        #[inline]
-        fn add(lhs: &Self::Field, rhs: &Self::Field, _: &mut Compiler<S>) -> Self::Field {
-            lhs + rhs
-        }
-
-        #[inline]
-        fn add_const(
-            lhs: &Self::Field,
-            rhs: &Self::ParameterField,
-            _: &mut Compiler<S>,
-        ) -> Self::Field {
-            lhs + FpVar::Constant(rhs.0)
-        }
-
-        #[inline]
-        fn mul(lhs: &Self::Field, rhs: &Self::Field, _: &mut Compiler<S>) -> Self::Field {
-            lhs * rhs
-        }
-
-        #[inline]
-        fn mul_const(
-            lhs: &Self::Field,
-            rhs: &Self::ParameterField,
-            _: &mut Compiler<S>,
-        ) -> Self::Field {
-            lhs * FpVar::Constant(rhs.0)
-        }
-
-        #[inline]
-        fn add_assign(lhs: &mut Self::Field, rhs: &Self::Field, _: &mut Compiler<S>) {
-            *lhs += rhs;
-        }
-
-        #[inline]
-        fn add_const_assign(
-            lhs: &mut Self::Field,
-            rhs: &Self::ParameterField,
-            _: &mut Compiler<S>,
-        ) {
-            *lhs += FpVar::Constant(rhs.0)
-        }
-
-        #[inline]
-        fn apply_sbox(point: &mut Self::Field, _: &mut Compiler<S>) {
-            *point = point
-                .pow_by_constant(&[Self::SBOX_EXPONENT])
-                .expect("Exponentiation is not allowed to fail.");
-        }
-    }
-
-    impl<S> Constant<Compiler<S>> for super::Permutation<S, Compiler<S>>
-    where
-        S: Specification,
-    {
-        type Type = super::Permutation<S>;
-
-        #[inline]
-        fn new_constant(this: &Self::Type, compiler: &mut Compiler<S>) -> Self {
-            let _ = compiler;
-            Self {
-                additive_round_keys: this.additive_round_keys.clone(),
-                mds_matrix: this.mds_matrix.clone(),
-            }
-        }
-    }
-}
-
-/// Testing Suite
-#[cfg(test)]
-mod test {
-    /// Tests if [`Poseidon2`](crate::config::Poseidon2) matches the known hash values.
-    #[test]
-    fn poseidon_hash_matches_known_values() {
-        /* TODO: After upgrading to new Poseidon, we have to enable these tests.
-        let hasher = Poseidon2::gen(&mut OsRng);
-        let inputs = [&Fp(field_new!(Fr, "1")), &Fp(field_new!(Fr, "2"))];
-        assert_eq!(
-            hasher.hash_untruncated(inputs, &mut ()),
-            vec![
-                Fp(field_new!(
-                    Fr,
-                    "1808609226548932412441401219270714120272118151392880709881321306315053574086"
-                )),
-                Fp(field_new!(
-                    Fr,
-                    "13469396364901763595452591099956641926259481376691266681656453586107981422876"
-                )),
-                Fp(field_new!(
-                    Fr,
-                    "28037046374767189790502007352434539884533225547205397602914398240898150312947"
-                )),
-            ]
-        );
-        */
     }
 }
